[RegistModal] rename : 파일 명 수정
