+++ src/components/Regist/RegistModalOne.jsx
@@ -1,5 +1,5 @@
 import { bool, string,func } from "prop-types";
-import ReactModal from "./RegistModal";
+import ReactModal from "./RegistModalBase";
 import styles from "./RegistModalOne.module.css";
 
 function RegistModalOne({ isOpen = false, onClose = false, title }) {
